# Pushdown dynamically generated filters into `MultiFileList`, allowing partitions to be pruned