--- a/src/common/multi_file_list.cpp
+++ b/src/common/multi_file_list.cpp
@@ -13,18 +13,21 @@
 
 namespace duckdb {
 
+MultiFilePushdownInfo::MultiFilePushdownInfo(LogicalGet &get) :
+	table_index(get.table_index), column_names(get.names), column_ids(get.column_ids), extra_info(get.extra_info) {}
+
 // Helper method to do Filter Pushdown into a MultiFileList
-bool PushdownInternal(ClientContext &context, const MultiFileReaderOptions &options, LogicalGet &get,
+bool PushdownInternal(ClientContext &context, const MultiFileReaderOptions &options, MultiFilePushdownInfo &info,
                       vector<unique_ptr<Expression>> &filters, vector<string> &expanded_files) {
 	unordered_map<string, column_t> column_map;
-	for (idx_t i = 0; i < get.column_ids.size(); i++) {
-		if (!IsRowIdColumnId(get.column_ids[i])) {
-			column_map.insert({get.names[get.column_ids[i]], i});
+	for (idx_t i = 0; i < info.column_ids.size(); i++) {
+		if (!IsRowIdColumnId(info.column_ids[i])) {
+			column_map.insert({info.column_names[info.column_ids[i]], i});
 		}
 	}
 
 	auto start_files = expanded_files.size();
-	HivePartitioning::ApplyFiltersToFileList(context, expanded_files, filters, column_map, get,
+	HivePartitioning::ApplyFiltersToFileList(context, expanded_files, filters, column_map, info,
 	                                         options.hive_partitioning, options.filename);
 
 	if (expanded_files.size() != start_files) {
@@ -124,7 +127,7 @@ bool MultiFileList::Scan(MultiFileListScanData &iterator, string &result_file) {
 }
 
 unique_ptr<MultiFileList> MultiFileList::ComplexFilterPushdown(ClientContext &context,
-                                                               const MultiFileReaderOptions &options, LogicalGet &get,
+                                                               const MultiFileReaderOptions &options, MultiFilePushdownInfo &info,
                                                                vector<unique_ptr<Expression>> &filters) {
 	// By default the filter pushdown into a multifilelist does nothing
 	return nullptr;
@@ -151,7 +154,7 @@ SimpleMultiFileList::SimpleMultiFileList(vector<string> paths_p)
 
 unique_ptr<MultiFileList> SimpleMultiFileList::ComplexFilterPushdown(ClientContext &context_p,
                                                                      const MultiFileReaderOptions &options,
-                                                                     LogicalGet &get,
+                                                                     MultiFilePushdownInfo &info,
                                                                      vector<unique_ptr<Expression>> &filters) {
 	if (!options.hive_partitioning && !options.filename) {
 		return nullptr;
@@ -159,7 +162,7 @@ unique_ptr<MultiFileList> SimpleMultiFileList::ComplexFilterPushdown(ClientConte
 
 	// FIXME: don't copy list until first file is filtered
 	auto file_copy = paths;
-	auto res = PushdownInternal(context_p, options, get, filters, file_copy);
+	auto res = PushdownInternal(context_p, options, info, filters, file_copy);
 
 	if (res) {
 		return make_uniq<SimpleMultiFileList>(file_copy);
@@ -203,7 +206,7 @@ GlobMultiFileList::GlobMultiFileList(ClientContext &context_p, vector<string> pa
 
 unique_ptr<MultiFileList> GlobMultiFileList::ComplexFilterPushdown(ClientContext &context_p,
                                                                    const MultiFileReaderOptions &options,
-                                                                   LogicalGet &get,
+                                                                   MultiFilePushdownInfo &info,
                                                                    vector<unique_ptr<Expression>> &filters) {
 	lock_guard<mutex> lck(lock);
 
@@ -216,7 +219,7 @@ unique_ptr<MultiFileList> GlobMultiFileList::ComplexFilterPushdown(ClientContext
 	if (!options.hive_partitioning && !options.filename) {
 		return nullptr;
 	}
-	auto res = PushdownInternal(context, options, get, filters, expanded_files);
+	auto res = PushdownInternal(context, options, info, filters, expanded_files);
 
 	if (res) {
 		return make_uniq<SimpleMultiFileList>(expanded_files);