--- a/src/common/multi_file_list.cpp
+++ b/src/common/multi_file_list.cpp
@@ -40,6 +40,29 @@ bool PushdownInternal(ClientContext &context, const MultiFileReaderOptions &opti
 	return false;
 }
 
+bool PushdownInternal(ClientContext &context, const MultiFileReaderOptions &options, const vector<string> &names,
+                                                        const vector<LogicalType> &types,
+                                                        const vector<column_t> &column_ids,
+                                                        const TableFilterSet &filters, vector<string> &expanded_files) {
+	idx_t table_index = 0;
+	ExtraOperatorInfo extra_info;
+
+	// construct the pushdown info
+	MultiFilePushdownInfo info(table_index, names, column_ids, extra_info);
+
+	// construct the set of expressions from the table filters
+	vector<unique_ptr<Expression>> filter_expressions;
+	for(auto &entry : filters.filters) {
+		auto column_idx = column_ids[entry.first];
+		auto column_ref = make_uniq<BoundColumnRefExpression>(types[column_idx], ColumnBinding(table_index, entry.first));
+		auto filter_expr = entry.second->ToExpression(*column_ref);
+		filter_expressions.push_back(std::move(filter_expr));
+	}
+
+	// call the original PushdownInternal method
+	return PushdownInternal(context, options, info, filter_expressions, expanded_files);
+}
+
 //===--------------------------------------------------------------------===//
 // MultiFileListIterator
 //===--------------------------------------------------------------------===//
@@ -136,6 +159,16 @@ unique_ptr<MultiFileList> MultiFileList::ComplexFilterPushdown(ClientContext &co
 	return nullptr;
 }
 
+unique_ptr<MultiFileList> MultiFileList::DynamicFilterPushdown(ClientContext &context,
+                                                const MultiFileReaderOptions &options,
+                                                const vector<string> &names,
+                                                const vector<LogicalType> &types,
+                                                const vector<column_t> &column_ids,
+                                                TableFilterSet &filters) const {
+	// By default the filter pushdown into a multifilelist does nothing
+	return nullptr;
+}
+
 unique_ptr<NodeStatistics> MultiFileList::GetCardinality(ClientContext &context) {
 	return nullptr;
 }
@@ -167,6 +200,26 @@ unique_ptr<MultiFileList> SimpleMultiFileList::ComplexFilterPushdown(ClientConte
 	auto file_copy = paths;
 	auto res = PushdownInternal(context_p, options, info, filters, file_copy);
 
+	if (res) {
+		return make_uniq<SimpleMultiFileList>(file_copy);
+	}
+
+	return nullptr;
+}
+
+unique_ptr<MultiFileList> SimpleMultiFileList::DynamicFilterPushdown(ClientContext &context,
+                                                        const MultiFileReaderOptions &options,
+                                                        const vector<string> &names,
+                                                        const vector<LogicalType> &types,
+                                                        const vector<column_t> &column_ids,
+                                                        TableFilterSet &filters) const {
+	if (!options.hive_partitioning && !options.filename) {
+		return nullptr;
+	}
+
+	// FIXME: don't copy list until first file is filtered
+	auto file_copy = paths;
+	auto res = PushdownInternal(context, options, names, types, column_ids, filters, file_copy);
 	if (res) {
 		return make_uniq<SimpleMultiFileList>(file_copy);
 	}
@@ -216,14 +269,13 @@ unique_ptr<MultiFileList> GlobMultiFileList::ComplexFilterPushdown(ClientContext
 	// Expand all
 	// FIXME: lazy expansion
 	// FIXME: push down filters into glob
-	while (ExpandPathInternal()) {
+	while (ExpandNextPath()) {
 	}
 
 	if (!options.hive_partitioning && !options.filename) {
 		return nullptr;
 	}
 	auto res = PushdownInternal(context, options, info, filters, expanded_files);
-
 	if (res) {
 		return make_uniq<SimpleMultiFileList>(expanded_files);
 	}
@@ -231,16 +283,43 @@ unique_ptr<MultiFileList> GlobMultiFileList::ComplexFilterPushdown(ClientContext
 	return nullptr;
 }
 
+unique_ptr<MultiFileList> GlobMultiFileList::DynamicFilterPushdown(ClientContext &context,
+                                                        const MultiFileReaderOptions &options,
+                                                        const vector<string> &names,
+                                                        const vector<LogicalType> &types,
+                                                        const vector<column_t> &column_ids,
+                                                        TableFilterSet &filters) const {
+	if (!options.hive_partitioning && !options.filename) {
+		return nullptr;
+	}
+	lock_guard<mutex> lck(lock);
+
+	// Expand all paths into a copy
+	// FIXME: lazy expansion and push filters into glob
+	idx_t path_index = current_path;
+	auto file_list = expanded_files;
+	while (ExpandPathInternal(path_index, file_list)) {
+	}
+
+	auto res = PushdownInternal(context, options, names, types, column_ids, filters, file_list);
+	if (res) {
+		return make_uniq<SimpleMultiFileList>(file_list);
+	}
+
+	return nullptr;
+}
+
+
 vector<string> GlobMultiFileList::GetAllFiles() {
 	lock_guard<mutex> lck(lock);
-	while (ExpandPathInternal()) {
+	while (ExpandNextPath()) {
 	}
 	return expanded_files;
 }
 
 idx_t GlobMultiFileList::GetTotalFileCount() {
 	lock_guard<mutex> lck(lock);
-	while (ExpandPathInternal()) {
+	while (ExpandNextPath()) {
 	}
 	return expanded_files.size();
 }
@@ -265,7 +344,7 @@ string GlobMultiFileList::GetFile(idx_t i) {
 
 string GlobMultiFileList::GetFileInternal(idx_t i) {
 	while (expanded_files.size() <= i) {
-		if (!ExpandPathInternal()) {
+		if (!ExpandNextPath()) {
 			return "";
 		}
 	}
@@ -273,22 +352,25 @@ string GlobMultiFileList::GetFileInternal(idx_t i) {
 	return expanded_files[i];
 }
 
-bool GlobMultiFileList::ExpandPathInternal() {
-	if (IsFullyExpanded()) {
+bool GlobMultiFileList::ExpandPathInternal(idx_t &current_path, vector<string> &result) const {
+	if (current_path >= paths.size()) {
 		return false;
 	}
 
 	auto &fs = FileSystem::GetFileSystem(context);
 	auto glob_files = fs.GlobFiles(paths[current_path], context, glob_options);
 	std::sort(glob_files.begin(), glob_files.end());
-	expanded_files.insert(expanded_files.end(), glob_files.begin(), glob_files.end());
+	result.insert(result.end(), glob_files.begin(), glob_files.end());
 
 	current_path++;
-
 	return true;
 }
 
-bool GlobMultiFileList::IsFullyExpanded() {
+bool GlobMultiFileList::ExpandNextPath() {
+	return ExpandPathInternal(current_path, expanded_files);
+}
+
+bool GlobMultiFileList::IsFullyExpanded() const {
 	return current_path == paths.size();
 }