--- a/src/include/duckdb/common/multi_file_list.hpp
+++ b/src/include/duckdb/common/multi_file_list.hpp
@@ -54,6 +54,7 @@ public:
 
 struct MultiFilePushdownInfo {
 	explicit MultiFilePushdownInfo(LogicalGet &get);
+	MultiFilePushdownInfo(idx_t table_index, const vector<string> &column_names, const vector<column_t> &column_ids, ExtraOperatorInfo &extra_info);
 
 	idx_t table_index;
 	const vector<string> &column_names;