… benchmark/tpch/join/partition_pushdown.benchmark
… benchmark/tpch/partition/partition_pushdown.benchmark
… benchmark/tpch/partition/partition_pushdown_csv.benchmark
… src/common/multi_file_reader.cpp
… src/function/table/read_csv.cpp
… src/include/duckdb/common/hive_partitioning.hpp
… src/include/duckdb/common/multi_file_reader.hpp
… src/include/duckdb/planner/expression.hpp
… src/include/duckdb/planner/expression/bound_aggregate_expression.hpp
… src/include/duckdb/planner/expression/bound_between_expression.hpp
… src/include/duckdb/planner/expression/bound_case_expression.hpp
… src/include/duckdb/planner/expression/bound_cast_expression.hpp
… src/include/duckdb/planner/expression/bound_columnref_expression.hpp
… src/include/duckdb/planner/expression/bound_comparison_expression.hpp
… src/include/duckdb/planner/expression/bound_conjunction_expression.hpp
… src/include/duckdb/planner/expression/bound_constant_expression.hpp
… src/include/duckdb/planner/expression/bound_default_expression.hpp
… src/include/duckdb/planner/expression/bound_expanded_expression.hpp
… src/include/duckdb/planner/expression/bound_function_expression.hpp
… src/include/duckdb/planner/expression/bound_lambda_expression.hpp
… src/include/duckdb/planner/expression/bound_lambdaref_expression.hpp
… src/include/duckdb/planner/expression/bound_operator_expression.hpp
… src/include/duckdb/planner/expression/bound_parameter_expression.hpp
… src/include/duckdb/planner/expression/bound_reference_expression.hpp
… src/include/duckdb/planner/expression/bound_subquery_expression.hpp
… src/include/duckdb/planner/expression/bound_unnest_expression.hpp
… src/include/duckdb/planner/expression/bound_window_expression.hpp
… src/include/duckdb/planner/filter/conjunction_filter.hpp
… src/include/duckdb/planner/filter/constant_filter.hpp
… src/include/duckdb/planner/filter/null_filter.hpp