@Mytherin 2024-07-11
`EXPLAIN` would not work, since the queries are generated dynamically at run-time they cannot be shown in `EXPLAIN` since that does not execute the query.

We could potentially show them during `EXPLAIN ANALYZE` - but we don't have any mechanisms for emitting extra operator info after the binding phase. That could be added (and would be useful) - but is out of scope for this PR.